Combination/requirements.txt
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pyyaml>=6.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
joblib>=1.3.0
lseg-data>=1.0.0
tqdm>=4.65.0
